tools/SchemaManager and tools/SchemaManager.Core namespaces fixed
