+++ test/SchemaManager.Core.UnitTests/SchemaManager.Core.UnitTests.csproj
@@ -17,7 +17,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\Microsoft.Health.SqlServer\Microsoft.Health.SqlServer.csproj" />
-    <ProjectReference Include="..\..\tools\SchemaManager.Core\SchemaManager.Core.csproj" />
+    <ProjectReference Include="..\..\tools\Microsoft.Health.SchemaManager.Core\Microsoft.Health.SchemaManager.Core.csproj" />
   </ItemGroup>
 
 </Project>
