+++ tools/Microsoft.Health.SchemaManager.Core/ISchemaManager.cs
@@ -7,10 +7,10 @@ using System;
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
+using Microsoft.Health.SchemaManager.Core.Model;
 using Microsoft.Health.SqlServer.Features.Schema.Manager.Model;
-using SchemaManager.Core.Model;
 
-namespace SchemaManager.Core
+namespace Microsoft.Health.SchemaManager.Core
 {
     public interface ISchemaManager
     {
