+++ tools/Microsoft.Health.SchemaManager.Core/Microsoft.Health.SchemaManager.Core.csproj
