+++ tools/Microsoft.Health.SchemaManager.Core/Model/MutuallyExclusiveType.cs
@@ -3,12 +3,14 @@
 // Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
 // -------------------------------------------------------------------------------------------------
 
-namespace SchemaManager
+namespace Microsoft.Health.SchemaManager.Core.Model
 {
-    public static class CommandNames
+    public class MutuallyExclusiveType
     {
-        public const string Apply = "apply";
-        public const string Available = "available";
-        public const string Current = "current";
+        public int Version { get; set; }
+
+        public bool Next { get; set; }
+
+        public bool Latest { get; set; }
     }
 }
