+++ tools/Microsoft.Health.SchemaManager.Core/Properties/AssemblyInfo.cs
