+++ tools/Microsoft.Health.SchemaManager.Core/Resources.resx
