+++ tools/Microsoft.Health.SchemaManager/GlobalSuppressions.cs
@@ -1,0 +1,10 @@
+﻿// -------------------------------------------------------------------------------------------------
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
+// -------------------------------------------------------------------------------------------------
+
+using System.Diagnostics.CodeAnalysis;
+
+[assembly: SuppressMessage("Naming", "CA1710:Identifiers should have correct suffix", Justification = "The collection suffix is not appropriate here", Scope = "type", Target = "~T:Microsoft.Health.SchemaManager.Commands.ApplyCommand")]
+[assembly: SuppressMessage("Naming", "CA1710:Identifiers should have correct suffix", Justification = "The collection suffix is not appropriate here", Scope = "type", Target = "~T:Microsoft.Health.SchemaManager.Commands.CurrentCommand")]
+[assembly: SuppressMessage("Naming", "CA1710:Identifiers should have correct suffix", Justification = "The collection suffix is not appropriate here", Scope = "type", Target = "~T:Microsoft.Health.SchemaManager.Commands.AvailableCommand")]
