+++ tools/Microsoft.Health.SchemaManager/Properties/AssemblyInfo.cs
