+++ tools/Microsoft.Health.SchemaManager/Properties/launchSettings.json
