+++ tools/Microsoft.Health.SchemaManager/Validators/RequiredOptionValidator.cs
@@ -7,7 +7,7 @@ using System.CommandLine;
 using System.Linq;
 using EnsureThat;
 
-namespace SchemaManager.Validators
+namespace Microsoft.Health.SchemaManager.Validators
 {
     public static class RequiredOptionValidator
     {
